Fix parser to return complete commands from minimal input
* Previously the parser would grab all comments *after* a complete
command. In a situation where input is buffered directly from stdin,
the parser would effectively wait until the start of the *next* command
(or EOF) to return the former
* Parser now will only try to grab a comment starting on the same line
as the last part of the command, so essentially it will read up to the
first newline after a command declaration
* Also tweaked case commands to not differentiate `pre_body_comments`,
since they are captured by the complete command itself
* Also tweaked `skip_whitespace` to only do two token lookahead (for
backslash-newline) only when necessary. Previously we would
unconditionally grab two tokens and check for backslash-newline,
however, if the first token was already a newline, we would end up
blocking until the next input or line from interactive input